Add userId to networth controller method

+++ apps/backend/src/main/net-worth/logic/delete-net-worth-service.ts
@@ -1,12 +1,22 @@
-import { Injectable } from "@nestjs/common";    
-import { PrismaService } from "../../../prisma/prisma.service";
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from '../../../prisma/prisma.service';
 
 @Injectable()
-export class DeleteNetWorthService{
-    constructor(private readonly prisma: PrismaService) {}
-    async execute(snapshotId: string) {
-        return this.prisma.netWorthSnapshot.delete({
-            where: { id: snapshotId },   
-        })
+export class DeleteNetWorthService {
+  constructor(private readonly prisma: PrismaService) {}
+  async execute(userId: string, snapshotId: string) {
+    const snapshot = this.prisma.netWorthSnapshot.findFirst({
+      where: {
+        userId: userId,
+        id: snapshotId,
+      },
+    });
+
+    if (!snapshot) {
+      throw new NotFoundException(' Net Worth snapshot not found');
     }
+    return this.prisma.netWorthSnapshot.delete({
+      where: { id: snapshotId },
+    });
+  }
 }
